% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aalen-Johansen.R
\name{aalen_johansen}
\alias{aalen_johansen}
\title{Compute the conditional Aalen-Johansen estimator.}
\usage{
aalen_johansen(
  data,
  x = NULL,
  a = NULL,
  k = NULL,
  p = NULL,
  collapse = FALSE,
  weights = NULL
)
}
\arguments{
\item{data}{A list of trajectory data for each individual.}

\item{x}{A numeric value for conditioning.}

\item{a}{A bandwidth. Defaults to Sheather and Jones selection wrt marginal.}

\item{k}{Number of nearest neighbours (alternative to kernel).}

\item{p}{An integer representing the number of states. The absorbing state is last.}

\item{collapse}{Logical, whether to collapse the last state of the model.}

\item{weights}{A vector of observation weights.}
}
\value{
A list containing the Aalen-Johansen estimator, the Nelson-Aalen estimator, and related quantities.
}
\description{
Compute the conditional Aalen-Johansen estimator.
}
