% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Additional_tools.R
\name{prodint}
\alias{prodint}
\title{Calculate the product integral of a matrix function}
\usage{
prodint(start, end, step_size, lambda)
}
\arguments{
\item{start}{Start time.}

\item{end}{End time.}

\item{step_size}{Step size of the grid.}

\item{lambda}{A given matrix function.}
}
\value{
The product integral of the given matrix function.
}
\description{
Calculate the product integral of a matrix function
}
